Rewind migrations
Because the publish date was set to `timezone.now()` instead of
`timezone.now`, every time the model was loaded it was creating a new
date value for the default. This was causing messages that a migration
was necessary. Each migration was just updating the default value to a
new, more recent date.

This commit removes these migrations. The next step will be to
autogenerate a single migration to fix the default publish date.